feat(observability): Sentry SDK across 3 binaries + hygiene unblockers
Adds sentry 0.47.0 (panic, tracing, reqwest, rustls, anyhow, backtrace,
contexts — default-features off, rustls only) and initialises it as
the first statement in each binary main:
  - lad                (src/main.rs)
  - llm-as-dom-mcp     (src/mcp_server/mod.rs)
  - lad-relay          (src/relay/main.rs)

Each binary gets an init_sentry() helper. Behaviour:
  - SENTRY_DSN unset or empty  -> full no-op, binary runs as before
  - SENTRY_DSN set             -> panics and tracing::error events flushed
                                  via layered subscriber
  - SENTRY_ENVIRONMENT         -> defaults to "production"
  - Release                    -> sentry::release_name! (cargo version)

Guard is bound to _sentry_guard so Drop flushes pending events at exit
without tripping unused_variables.

Hygiene fixes required to unblock cargo check / clippy / build:
  - Cargo.toml   dedup duplicate url key (kept 2.5.8 pin)
  - Cargo.lock   regenerated (corrupt with duplicate fallible-iterator)
  - qr.rs        code.width() -> code.width() (qrcode 0.14 API change,
                 clippy -D warnings was failing on cast)

Incident response 2026-04-03: Sentry gives operational visibility so
future issues surface before turning into incidents like the npm leak.

cargo check       ok (5.56s)
cargo clippy -D   ok (0.54s)
cargo build --rel ok (56.18s)
SENTRY_DSN= run   identical behaviour, zero sentry noise

+++ src/relay/qr.rs
@@ -26,7 +26,7 @@ pub fn print_qr_stderr(url: &str) {
     };
 
     let modules = code.to_colors();
-    let width = code.width() as usize;
+    let width = code.width();
 
     // Quiet zone: 2 modules on each side.
     let qz = 2;
